+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### Changes
 
 ### Fixes
+- [#2373](https://gitlab.com/meltano/meltano/-/issues/2373) `meltano init` emits warning instead of failing when underlying filesystem doesn't support symlinks 
 
 ### Breaks